Fix and re-enable Ubuntu 12.x builds in Travis file Add function `rawrtc_dns_type_to_address_family_name` to utils Move ICE server functions into their own file Move ICE gather options functions into their own file
