+++ src/librawrtc/ice_gather_options.h
@@ -1,0 +1,5 @@
+#pragma once
+
+enum rawrtc_code rawrtc_ice_gather_options_destroy_url_dns_contexts(
+    struct rawrtc_ice_gather_options* const options // not checked
+);
